bots/pocoyo/requirements.txt
python-telegram-bot>=20.7,<22
